cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "axon-cli"
version = "0.1.0"
description = "Sync Claude configuration from Axon platform"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.27.0",
    "keyring>=25.0.0",
    "rich>=14.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-httpx>=0.30.0",
]

[project.scripts]
axon = "atlas_cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/atlas_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
